# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VENDORA API",
        "description": "API documentation for VENDORA.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application base URL"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "List customers",
                "operationId": "1b873d7b0c63876f30a978b52c4d24d2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Dela Cruz"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john.delacruz@email.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+63 912 345 6789"
                                                    },
                                                    "orders_count": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "total_spent": {
                                                        "type": "integer",
                                                        "example": 15420
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-01-10T10:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-01-10T10:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 342
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create a customer",
                "operationId": "747630c757eae4c38404384330e68c57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Dela Cruz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.delacruz@email.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+63 912 345 6789"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Dela Cruz"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john.delacruz@email.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+63 912 345 6789"
                                        },
                                        "orders_count": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total_spent": {
                                            "type": "integer",
                                            "example": 15420
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-01-10T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-01-10T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customers/summary": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer summary cards",
                "operationId": "86a8d6debd830715555e8ba17774f19c",
                "responses": {
                    "200": {
                        "description": "Customer summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_customers": {
                                            "type": "integer",
                                            "example": 342
                                        },
                                        "active": {
                                            "type": "integer",
                                            "example": 298
                                        },
                                        "vip": {
                                            "type": "integer",
                                            "example": 24
                                        },
                                        "inactive": {
                                            "type": "integer",
                                            "example": 20
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/customers/{customer}/credits": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get credit history for a customer",
                "operationId": "126faf68401126fc9693a583bf3fc273",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credit history"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}